+++ src/fp/parseDateWithOptions/index.js
@@ -1,0 +1,8 @@
+// This file is generated automatically by `scripts/build/fp.js`. Please, don't change it.
+
+import fn from '../../parseDate/index.js'
+import convertToFP from '../_lib/convertToFP/index.js'
+
+var parseDateWithOptions = convertToFP(fn, 4)
+
+export default parseDateWithOptions