+++ src/parseDate/index.d.ts
@@ -1,0 +1,4 @@
+// This file is generated automatically by `scripts/build/typings.js`. Please, don't change it.
+
+import {parseDate} from 'date-fns'
+export = parseDate