+++ typings.d.ts
@@ -757,6 +757,14 @@ declare module 'date-fns' {
   ): Date
   namespace parse {}
 
+  function parseDate (
+    dateString: string,
+    formatString: string,
+    baseDate: Date | string | number,
+    options?: Options
+  ): Date
+  namespace parseDate {}
+
   function setDate (
     date: Date | string | number,
     dayOfMonth: number,
@@ -1530,6 +1538,11 @@ declare module 'date-fns/parse' {
   export = parse
 }
 
+declare module 'date-fns/parseDate' {
+  import {parseDate} from 'date-fns'
+  export = parseDate
+}
+
 declare module 'date-fns/setDate' {
   import {setDate} from 'date-fns'
   export = setDate
@@ -2240,6 +2253,11 @@ declare module 'date-fns/parse/index' {
   export = parse
 }
 
+declare module 'date-fns/parseDate/index' {
+  import {parseDate} from 'date-fns'
+  export = parseDate
+}
+
 declare module 'date-fns/setDate/index' {
   import {setDate} from 'date-fns'
   export = setDate
@@ -2948,6 +2966,11 @@ declare module 'date-fns/min/index.js' {
 declare module 'date-fns/parse/index.js' {
   import {parse} from 'date-fns'
   export = parse
+}
+
+declare module 'date-fns/parseDate/index.js' {
+  import {parseDate} from 'date-fns'
+  export = parseDate
 }
 
 declare module 'date-fns/setDate/index.js' {
@@ -3764,6 +3787,12 @@ declare module 'date-fns/fp' {
 
   const parse: CurriedFn3<Date | string | number, string, string, Date>
   namespace parse {}
+
+  const parseDate: CurriedFn3<Date | string | number, string, string, Date>
+  namespace parseDate {}
+
+  const parseDateWithOptions: CurriedFn4<Options, Date | string | number, string, string, Date>
+  namespace parseDateWithOptions {}
 
   const parseWithOptions: CurriedFn4<Options, Date | string | number, string, string, Date>
   namespace parseWithOptions {}
@@ -5034,6 +5063,16 @@ declare module 'date-fns/fp/minWithOptions' {
 declare module 'date-fns/fp/parse' {
   import {parse} from 'date-fns/fp'
   export = parse
+}
+
+declare module 'date-fns/fp/parseDate' {
+  import {parseDate} from 'date-fns/fp'
+  export = parseDate
+}
+
+declare module 'date-fns/fp/parseDateWithOptions' {
+  import {parseDateWithOptions} from 'date-fns/fp'
+  export = parseDateWithOptions
 }
 
 declare module 'date-fns/fp/parseWithOptions' {
@@ -6456,6 +6495,16 @@ declare module 'date-fns/fp/parse/index' {
   export = parse
 }
 
+declare module 'date-fns/fp/parseDate/index' {
+  import {parseDate} from 'date-fns/fp'
+  export = parseDate
+}
+
+declare module 'date-fns/fp/parseDateWithOptions/index' {
+  import {parseDateWithOptions} from 'date-fns/fp'
+  export = parseDateWithOptions
+}
+
 declare module 'date-fns/fp/parseWithOptions/index' {
   import {parseWithOptions} from 'date-fns/fp'
   export = parseWithOptions
@@ -7876,6 +7925,16 @@ declare module 'date-fns/fp/parse/index.js' {
   export = parse
 }
 
+declare module 'date-fns/fp/parseDate/index.js' {
+  import {parseDate} from 'date-fns/fp'
+  export = parseDate
+}
+
+declare module 'date-fns/fp/parseDateWithOptions/index.js' {
+  import {parseDateWithOptions} from 'date-fns/fp'
+  export = parseDateWithOptions
+}
+
 declare module 'date-fns/fp/parseWithOptions/index.js' {
   import {parseWithOptions} from 'date-fns/fp'
   export = parseWithOptions
@@ -8937,6 +8996,14 @@ declare module 'date-fns/esm' {
   ): Date
   namespace parse {}
 
+  function parseDate (
+    dateString: string,
+    formatString: string,
+    baseDate: Date | string | number,
+    options?: Options
+  ): Date
+  namespace parseDate {}
+
   function setDate (
     date: Date | string | number,
     dayOfMonth: number,
@@ -9710,6 +9777,11 @@ declare module 'date-fns/esm/parse' {
   export default parse
 }
 
+declare module 'date-fns/esm/parseDate' {
+  import {parseDate} from 'date-fns/esm'
+  export default parseDate
+}
+
 declare module 'date-fns/esm/setDate' {
   import {setDate} from 'date-fns/esm'
   export default setDate
@@ -10420,6 +10492,11 @@ declare module 'date-fns/esm/parse/index' {
   export default parse
 }
 
+declare module 'date-fns/esm/parseDate/index' {
+  import {parseDate} from 'date-fns/esm'
+  export default parseDate
+}
+
 declare module 'date-fns/esm/setDate/index' {
   import {setDate} from 'date-fns/esm'
   export default setDate
@@ -11128,6 +11205,11 @@ declare module 'date-fns/esm/min/index.js' {
 declare module 'date-fns/esm/parse/index.js' {
   import {parse} from 'date-fns/esm'
   export default parse
+}
+
+declare module 'date-fns/esm/parseDate/index.js' {
+  import {parseDate} from 'date-fns/esm'
+  export default parseDate
 }
 
 declare module 'date-fns/esm/setDate/index.js' {
@@ -11944,6 +12026,12 @@ declare module 'date-fns/esm/fp' {
 
   const parse: CurriedFn3<Date | string | number, string, string, Date>
   namespace parse {}
+
+  const parseDate: CurriedFn3<Date | string | number, string, string, Date>
+  namespace parseDate {}
+
+  const parseDateWithOptions: CurriedFn4<Options, Date | string | number, string, string, Date>
+  namespace parseDateWithOptions {}
 
   const parseWithOptions: CurriedFn4<Options, Date | string | number, string, string, Date>
   namespace parseWithOptions {}
@@ -13214,6 +13302,16 @@ declare module 'date-fns/esm/fp/minWithOptions' {
 declare module 'date-fns/esm/fp/parse' {
   import {parse} from 'date-fns/esm/fp'
   export default parse
+}
+
+declare module 'date-fns/esm/fp/parseDate' {
+  import {parseDate} from 'date-fns/esm/fp'
+  export default parseDate
+}
+
+declare module 'date-fns/esm/fp/parseDateWithOptions' {
+  import {parseDateWithOptions} from 'date-fns/esm/fp'
+  export default parseDateWithOptions
 }
 
 declare module 'date-fns/esm/fp/parseWithOptions' {
@@ -14636,6 +14734,16 @@ declare module 'date-fns/esm/fp/parse/index' {
   export default parse
 }
 
+declare module 'date-fns/esm/fp/parseDate/index' {
+  import {parseDate} from 'date-fns/esm/fp'
+  export default parseDate
+}
+
+declare module 'date-fns/esm/fp/parseDateWithOptions/index' {
+  import {parseDateWithOptions} from 'date-fns/esm/fp'
+  export default parseDateWithOptions
+}
+
 declare module 'date-fns/esm/fp/parseWithOptions/index' {
   import {parseWithOptions} from 'date-fns/esm/fp'
   export default parseWithOptions
@@ -16054,6 +16162,16 @@ declare module 'date-fns/esm/fp/minWithOptions/index.js' {
 declare module 'date-fns/esm/fp/parse/index.js' {
   import {parse} from 'date-fns/esm/fp'
   export default parse
+}
+
+declare module 'date-fns/esm/fp/parseDate/index.js' {
+  import {parseDate} from 'date-fns/esm/fp'
+  export default parseDate
+}
+
+declare module 'date-fns/esm/fp/parseDateWithOptions/index.js' {
+  import {parseDateWithOptions} from 'date-fns/esm/fp'
+  export default parseDateWithOptions
 }
 
 declare module 'date-fns/esm/fp/parseWithOptions/index.js' {
@@ -18670,6 +18788,13 @@ interface dateFns {
   ): Date
 
   parse(
+    dateString: string,
+    formatString: string,
+    baseDate: Date | string | number,
+    options?: Options
+  ): Date
+
+  parseDate(
     dateString: string,
     formatString: string,
     baseDate: Date | string | number,